requirements.txt
asgiref==3.5.2
autopep8==1.7.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
pycodestyle==2.9.1
pytz==2022.5
sqlparse==0.4.3
tomli==2.0.1
